ai-image-generator/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
diffusers
transformers
accelerate
xformers
safetensors
websockets
python-multipart
pydantic
python-dotenv
